requirements.txt
attrs==24.2.0
autobahn==19.11.2
Automat==22.10.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chromedriver-binary==129.0.6659.0.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiteconnect==5.0.1
MarkupSafe==2.1.5
outcome==1.3.0.post0
panda==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
pyotp==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.2
requests==2.32.3
selenium==4.23.1
service-identity==24.1.0
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
zope.interface==7.0.1
